+++ app/views/original_books/_book_list.html.slim
@@ -39,10 +39,13 @@
           td
             = book.isbn
           td
-            = link_to '積ん読に追加する', original_books_path(original_book:{author: book.author, title: book.title, publisher: book.publisher, publish_date: book.publication_date, isbn: book.isbn, image_url: book.image_url, url: book.url}, status:"tsundoku"), method: :post
+            = link_to '積ん読に追加する', original_books_path(set_book(book,"tsundoku")), method: :post
           td
-            = link_to "読んだ本に追加する", original_books_path(original_book:{author: book.author, title: book.title, publisher: book.publisher, publish_date: book.publication_date, isbn: book.isbn, image_url: book.image_url, url: book.url}, status:"read"), method: :post
+            = link_to "読んだ本に追加する", original_books_path(set_book(book,"read")), method: :post
           td
-            = link_to '読みたい本に追加する', original_books_path(original_book:{author: book.author, title: book.title, publisher: book.publisher, publish_date: book.publication_date, isbn: book.isbn, image_url: book.image_url, url: book.url}, status:"wish"), method: :post
+            = link_to '読みたい本に追加する', original_books_path(set_book(book,"wish")), method: :post
           /td
           /  = link_to 'reviewを書く', new_reviews_path()
+          / shelved_book/:id/review
+          / =>shelved_book.params[:id]
+          / サービスクラス、デコレータークラスに切り出して行くというのもあります。
