+++ config/routes.rb
@@ -7,7 +7,11 @@ Rails.application.routes.draw do
     omniauth_callbacks: "users/omniauth_callbacks"
   }
 
-  resources :users, only: [:index, :show, :edit, :update]
+  resources :users, only: [:index, :show, :edit, :update] do
+  collection do
+      get '/:id/shelf', to: 'users#shelf', as: 'shelf'
+    end
+  end
   resources :reviews
   resources :original_books do
   collection do
@@ -15,11 +19,7 @@ Rails.application.routes.draw do
     end
   end
 
-  resources :shelved_books do
-  collection do
-      get '/:user_id/shelf', to: 'shelved_books#shelf'
-    end
-  end
+  resources :shelved_books
 
   devise_scope :user do
     authenticated :user do
